# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE GUIDE International Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResult {
  success: boolean;
  message?: string;
}

export interface Course {
  id: number;
  title: string;
  slug: string;
  category: string;
  duration: string;
  description: string;
  /** @nullable */
  eligibility?: string | null;
  /** @nullable */
  syllabus?: string | null;
  /** @nullable */
  careerOpportunities?: string | null;
  /** @nullable */
  certificationDetails?: string | null;
  /** @nullable */
  benefits?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  sortOrder?: number;
  createdAt?: string;
}

export interface CourseInput {
  title: string;
  slug: string;
  category: string;
  duration: string;
  description: string;
  eligibility?: string;
  syllabus?: string;
  careerOpportunities?: string;
  certificationDetails?: string;
  benefits?: string;
  imageUrl?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface CourseUpdate {
  title?: string;
  slug?: string;
  category?: string;
  duration?: string;
  description?: string;
  eligibility?: string;
  syllabus?: string;
  careerOpportunities?: string;
  certificationDetails?: string;
  benefits?: string;
  imageUrl?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface Student {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  certificateNumber: string;
  courseName: string;
  completionDate: string;
  grade: string;
  status: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt?: string;
}

export interface StudentInput {
  name: string;
  photoUrl?: string;
  certificateNumber: string;
  courseName: string;
  completionDate: string;
  grade: string;
  status?: string;
  email?: string;
  phone?: string;
}

export interface StudentUpdate {
  name?: string;
  photoUrl?: string;
  certificateNumber?: string;
  courseName?: string;
  completionDate?: string;
  grade?: string;
  status?: string;
  email?: string;
  phone?: string;
}

export interface StudentsPage {
  data: Student[];
  total: number;
  page: number;
  limit: number;
}

export interface BulkImportInput {
  records: StudentInput[];
}

export interface BulkImportResult {
  imported: number;
  failed: number;
  total: number;
  errors?: string[];
}

export interface Certificate {
  id: number;
  certificateNumber: string;
  studentName: string;
  /** @nullable */
  studentPhotoUrl?: string | null;
  courseName: string;
  completionDate: string;
  grade: string;
  status: string;
  /** @nullable */
  studentId?: number | null;
  createdAt?: string;
}

export interface CertificateInput {
  certificateNumber: string;
  studentName: string;
  studentPhotoUrl?: string;
  courseName: string;
  completionDate: string;
  grade: string;
  status?: string;
  studentId?: number;
}

export interface CertificateUpdate {
  certificateNumber?: string;
  studentName?: string;
  studentPhotoUrl?: string;
  courseName?: string;
  completionDate?: string;
  grade?: string;
  status?: string;
}

export interface CertificatesPage {
  data: Certificate[];
  total: number;
  page: number;
  limit: number;
}

export interface VerificationResult {
  found: boolean;
  certificate?: Certificate;
  verifiedAt?: string;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  category: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
  createdAt?: string;
}

export interface GalleryImageInput {
  imageUrl: string;
  category: string;
  title?: string;
  description?: string;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  studentName: string;
  /** @nullable */
  photoUrl?: string | null;
  courseName: string;
  rating: number;
  review: string;
  /** @nullable */
  designation?: string | null;
  approved: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  studentName: string;
  photoUrl?: string;
  courseName: string;
  rating: number;
  review: string;
  designation?: string;
  approved?: boolean;
}

export interface TestimonialUpdate {
  studentName?: string;
  photoUrl?: string;
  courseName?: string;
  rating?: number;
  review?: string;
  designation?: string;
  approved?: boolean;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  /** @nullable */
  courseInterest?: string | null;
  status: string;
  createdAt?: string;
}

export interface InquiryInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  courseInterest?: string;
}

export interface ResolveInput {
  status: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message?: string;
}

export interface AdminInfo {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
}

export interface DashboardStats {
  totalStudents: number;
  totalCertificates: number;
  totalCourses: number;
  totalInquiries: number;
  pendingInquiries: number;
  recentVerificationsCount: number;
}

export interface VerificationLog {
  id: number;
  certificateNumber: string;
  found: boolean;
  /** @nullable */
  ipAddress?: string | null;
  verifiedAt: string;
}

export interface SiteSettings {
  id: number;
  instituteName?: string;
  tagline?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  youtube?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  heroBannerUrl?: string | null;
}

export interface SiteSettingsUpdate {
  instituteName?: string;
  tagline?: string;
  logoUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  whatsapp?: string;
  facebook?: string;
  instagram?: string;
  linkedin?: string;
  youtube?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  heroBannerUrl?: string;
}

export type ListCoursesParams = {
category?: string;
featured?: boolean;
};

export type ListStudentsParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListCertificatesParams = {
search?: string;
page?: number;
limit?: number;
};

export type VerifyCertificateParams = {
number: string;
};

export type ListGalleryParams = {
category?: string;
};

export type ListTestimonialsParams = {
approved?: boolean;
};

export type ListInquiriesParams = {
status?: string;
};