# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THE GUIDE International Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Students ──────────────────────────────────────────────────
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List all certificates
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Certificates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificatesPage"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Create a certificate record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/verify:
    get:
      operationId: verifyCertificate
      tags: [certificates]
      summary: Verify a certificate by number
      parameters:
        - name: number
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /certificates/{id}:
    patch:
      operationId: updateCertificate
      tags: [certificates]
      summary: Update a certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
    delete:
      operationId: deleteCertificate
      tags: [certificates]
      summary: Delete a certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Gallery ──────────────────────────────────────────────────
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add a gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete a gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Testimonials ──────────────────────────────────────────────
...
  /contact:
    get:
      operationId: listInquiries
      tags: [contact]
      summary: List contact inquiries
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [contact]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /contact/{id}/resolve:
    patch:
      operationId: resolveInquiry
      tags: [contact]
      summary: Mark inquiry as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /contact/{id}:
    delete:
      operationId: deleteInquiry
      tags: [contact]
      summary: Delete an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── Admin Auth ────────────────────────────────────────────────
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ── Course ──────────────────────────────────────────────────
    Course:
      type: object
      required: [id, title, slug, category, duration, description]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        duration:
          type: string
        description:
          type: string
        eligibility:
          type: ["string", "null"]
        syllabus:
          type: ["string", "null"]
        careerOpportunities:
          type: ["string", "null"]
        certificationDetails:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    CourseInput:
      type: object
      required: [title, slug, category, duration, description]
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        duration:
          type: string
        description:
          type: string
        eligibility:
          type: string
        syllabus:
          type: string
        careerOpportunities:
          type: string
        certificationDetails:
          type: string
        benefits:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        duration:
          type: string
        description:
          type: string
        eligibility:
          type: string
        syllabus:
          type: string
        careerOpportunities:
          type: string
        certificationDetails:
          type: string
        benefits:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer

    # ── Student ──────────────────────────────────────────────────
...
    Certificate:
      type: object
      required: [id, certificateNumber, studentName, courseName, completionDate, grade, status]
      properties:
        id:
          type: integer
        certificateNumber:
          type: string
        studentName:
          type: string
        studentPhotoUrl:
          type: ["string", "null"]
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
        studentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    CertificateInput:
      type: object
      required: [certificateNumber, studentName, courseName, completionDate, grade]
      properties:
        certificateNumber:
          type: string
        studentName:
          type: string
        studentPhotoUrl:
          type: string
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
        studentId:
          type: integer
    CertificateUpdate:
      type: object
      properties:
        certificateNumber:
          type: string
        studentName:
          type: string
        studentPhotoUrl:
          type: string
        courseName:
          type: string
        completionDate:
          type: string
        grade:
          type: string
        status:
          type: string
    CertificatesPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    VerificationResult:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        certificate:
          $ref: "#/components/schemas/Certificate"
        verifiedAt:
          type: string

    # ── Gallery ──────────────────────────────────────────────────
    GalleryImage:
      type: object
      required: [id, imageUrl, category]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    GalleryImageInput:
      type: object
      required: [imageUrl, category]
      properties:
        imageUrl:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer

    # ── Testimonial ──────────────────────────────────────────────
...
    Inquiry:
      type: object
      required: [id, name, email, subject, message, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        courseInterest:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    InquiryInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        courseInterest:
          type: string
    ResolveInput:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ── Admin ──────────────────────────────────────────────────
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...